[po-199] Move campaignCollectiveResult router into campaign
