+++ api/lib/routes.js
@@ -4,7 +4,6 @@ module.exports = [
   require('./application/assessments'),
   require('./application/authentication'),
   require('./application/cache'),
-  require('./application/campaign-collective-results'),
   require('./application/campaignParticipationResults'),
   require('./application/campaignParticipations'),
   require('./application/campaignReports'),
